Refactor webpack config and update plugins.
Sync plugins version according to ReactTemplate: https://github.com/genesis-dean/ReactTemplate
