+++ package.json
@@ -4,9 +4,10 @@
 	"description": "",
 	"main": "index.js",
 	"scripts": {
-		"start": "webpack serve --config ./config/webpack.dev.config.ts --open",
-		"build": "webpack --config ./config/webpack.prod.config.ts",
-		"test": "echo \"Error: no test specified\" && exit 1"
+		"start": "webpack serve --config ./config/webpack.config.dev.ts --open",
+		"build": "webpack --config ./config/webpack.config.prod.ts",
+		"test": "echo \"Error: no test specified\" && exit 1",
+		"lint": "./node_modules/.bin/eslint ./src/**/*"
 	},
 	"author": "",
 	"license": "ISC",
@@ -20,29 +21,29 @@
 		"react-router-dom": "^5.2.0"
 	},
 	"devDependencies": {
-		"@babel/core": "^7.14.3",
-		"@babel/plugin-transform-runtime": "^7.14.3",
-		"@babel/preset-env": "^7.14.4",
-		"@babel/preset-react": "^7.13.13",
-		"@babel/preset-typescript": "^7.13.0",
-		"@babel/runtime": "^7.14.0",
+		"@babel/core": "^7.14.6",
+		"@babel/plugin-transform-runtime": "^7.14.5",
+		"@babel/preset-env": "^7.14.7",
+		"@babel/preset-react": "^7.14.5",
+		"@babel/preset-typescript": "^7.14.5",
+		"@babel/runtime": "^7.14.6",
 		"@types/fork-ts-checker-webpack-plugin": "^0.4.5",
 		"@types/node": "^15.12.4",
-		"@types/react": "^17.0.8",
-		"@types/react-dom": "^17.0.5",
+		"@types/react": "^17.0.14",
+		"@types/react-dom": "^17.0.9",
 		"@types/webpack": "^5.28.0",
-		"@types/webpack-dev-server": "^3.11.4",
-		"@typescript-eslint/eslint-plugin": "^4.25.0",
-		"@typescript-eslint/parser": "^4.25.0",
+		"@types/webpack-dev-server": "^3.11.5",
+		"@typescript-eslint/eslint-plugin": "^4.28.2",
+		"@typescript-eslint/parser": "^4.28.2",
 		"babel-loader": "^8.2.2",
 		"clean-webpack-plugin": "^4.0.0-alpha.0",
 		"css-loader": "^6.0.0",
-		"eslint": "^7.27.0",
+		"eslint": "^7.30.0",
 		"eslint-plugin-react": "^7.24.0",
 		"eslint-plugin-react-hooks": "^4.2.0",
 		"eslint-webpack-plugin": "^2.5.4",
-		"fork-ts-checker-webpack-plugin": "^6.2.10",
-		"html-webpack-plugin": "^5.3.1",
+		"fork-ts-checker-webpack-plugin": "^6.2.12",
+		"html-webpack-plugin": "^5.3.2",
 		"node-polyfill-webpack-plugin": "^1.1.4",
 		"postcss": "^8.3.5",
 		"postcss-loader": "^6.1.0",
@@ -50,9 +51,10 @@
 		"sass-loader": "^12.1.0",
 		"style-loader": "^3.1.0",
 		"ts-node": "^10.0.0",
-		"typescript": "^4.3.2",
-		"webpack": "^5.40.0",
-		"webpack-cli": "^4.7.0",
-		"webpack-dev-server": "^3.11.2"
+		"typescript": "^4.3.5",
+		"webpack": "^5.43.0",
+		"webpack-cli": "^4.7.2",
+		"webpack-dev-server": "^3.11.2",
+		"webpack-merge": "^5.8.0"
 	}
 }
